requirements.txt
django==4.2.1
pytest==7.3.1
pytest-django==4.5.2
requests==2.30.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0